split out last run from upload and update test
