+++ app/src/organisms/RunTimeControl/__tests__/Timer.test.tsx
@@ -3,7 +3,7 @@ import { renderWithProviders } from '@opentrons/components'
 
 import { i18n } from '../../../i18n'
 import { Timer } from '../Timer'
-import { useRunCompleteTime, useRunPauseTime, useRunStopTime } from '../hooks'
+import { useRunTimestamps } from '../hooks'
 import {
   RUN_STATUS_RUNNING,
   RUN_STATUS_STOP_REQUESTED,
@@ -16,14 +16,8 @@ const PAUSED_TIME = '2021-10-07T18:47:55.366581+00:00'
 const COMPLETED_TIME = '2021-10-07T18:58:59.366581+00:00'
 const STOPPED_TIME = '2021-10-07T18:45:49.366581+00:00'
 
-const mockUseRunCompleteTime = useRunCompleteTime as jest.MockedFunction<
-  typeof useRunCompleteTime
->
-const mockUseRunStopTime = useRunStopTime as jest.MockedFunction<
-  typeof useRunStopTime
->
-const mockUseRunPauseTime = useRunPauseTime as jest.MockedFunction<
-  typeof useRunPauseTime
+const mockUseRunTimestamps = useRunTimestamps as jest.MockedFunction<
+  typeof useRunTimestamps
 >
 
 describe('Timer', () => {
@@ -32,9 +26,12 @@ describe('Timer', () => {
   ) => ReturnType<typeof renderWithProviders>
 
   beforeEach(() => {
-    mockUseRunCompleteTime.mockReturnValue(null)
-    mockUseRunPauseTime.mockReturnValue(null)
-    mockUseRunStopTime.mockReturnValue(null)
+    mockUseRunTimestamps.mockReturnValue({
+      startedAt: null,
+      pausedAt: null,
+      stoppedAt: null,
+      completedAt: null,
+    })
     render = (
       props = {
         startTime: START_TIME,
@@ -69,7 +66,12 @@ describe('Timer', () => {
   })
 
   it('renders a paused time and a run time when paused', () => {
-    mockUseRunPauseTime.mockReturnValue(PAUSED_TIME)
+    mockUseRunTimestamps.mockReturnValue({
+      startedAt: null,
+      pausedAt: PAUSED_TIME,
+      stoppedAt: null,
+      completedAt: null,
+    })
     const [{ getAllByText, getByText }] = render()
 
     expect(getByText('Run Time:')).toBeTruthy()
@@ -78,7 +80,12 @@ describe('Timer', () => {
   })
 
   it('renders a completed time when completed', async () => {
-    mockUseRunCompleteTime.mockReturnValue(COMPLETED_TIME)
+    mockUseRunTimestamps.mockReturnValue({
+      startedAt: null,
+      pausedAt: null,
+      stoppedAt: null,
+      completedAt: COMPLETED_TIME,
+    })
     const [{ getByText }] = render()
 
     expect(getByText('Run Time:')).toBeTruthy()
@@ -86,7 +93,12 @@ describe('Timer', () => {
   })
 
   it('renders a stopped time when run is canceled', () => {
-    mockUseRunStopTime.mockReturnValue(STOPPED_TIME)
+    mockUseRunTimestamps.mockReturnValue({
+      startedAt: null,
+      pausedAt: null,
+      stoppedAt: STOPPED_TIME,
+      completedAt: null,
+    })
     const [{ getByText }] = render({
       startTime: START_TIME,
       runStatus: RUN_STATUS_STOP_REQUESTED,
